fix: move `RealisticCallParams` type to its own file and update compilation guard. (#6602)
fix: move `RealisticCallParams` type to its own file and also allow compilation for doc tests.
